+++ Core/HoverUIData.cs
@@ -24,7 +24,7 @@ namespace Radiance.Core
         public HoverUIData parent;
         public Vector2 elementPosition;
         public Vector2 targetPosition;
-        public const float timerMax = 20;
+        public const float TIMER_MAX = 20;
         public float timer;
         public string name;
         public bool updateTimer = false;
