+++ Core/RadianceGlobalItem.cs
@@ -9,26 +9,6 @@ namespace Radiance.Core
         public override bool InstancePerEntity => true;
         public int formationPickupTimer = 0;
 
-        public override void Load()
-        {
-            TransmutatorTileEntity.PostTransmutateItemEvent += GiveFishUnlock;
-        }
-
-        private void GiveFishUnlock(TransmutatorTileEntity transmutator, TransmutationRecipe recipe)
-        {
-            if (RadianceSets.ProjectorLensID[transmutator.projector.LensPlaced.type] == (int)ProjectorLensID.Fish)
-                UnlockSystem.transmutatorFishUsed = true;
-        }
-
-        public override void SetStaticDefaults()
-        {
-            RadianceSets.ProjectorLensTexture[ItemID.SpecularFish] = "Radiance/Content/Tiles/Transmutator/SpecularFish_Transmutator";
-            RadianceSets.ProjectorLensID[ItemID.SpecularFish] = (int)ProjectorLensID.Fish;
-            RadianceSets.ProjectorLensDust[ItemID.SpecularFish] = DustID.FrostDaggerfish;
-            RadianceSets.ProjectorLensSound[ItemID.SpecularFish] = new SoundStyle($"{nameof(Radiance)}/Sounds/FishSplat");
-            RadianceSets.ProjectorLensPreOrderedUpdateFunction[ItemID.SpecularFish] = (projector) => projector.transmutator.radianceModifier *= 25f;
-        }
-
         public override bool PreDrawInInventory(Item item, SpriteBatch spriteBatch, Vector2 position, Rectangle frame, Color drawColor, Color itemColor, Vector2 origin, float scale)
         {
             if (Main.LocalPlayer.GetModPlayer<RadianceInterfacePlayer>().canSeeLensItems && RadianceSets.ProjectorLensID[item.type] != -1)
