+++ Core/RadianceSets.cs
@@ -8,6 +8,9 @@ namespace Radiance.Core
         public static SetFactory TilesFactory = new SetFactory(TileLoader.TileCount);
 
         public static string[] EncycloradiaRelatedEntry = ItemFactory.CreateCustomSet(string.Empty);
+        public static Func<Item[], int, bool>[] RightClickMouseItemFunction = ItemFactory.CreateCustomSet<Func<Item[], int, bool>>(null);
+        public static float[] SetPedestalStability = ItemFactory.CreateFloatSet(0);
+        public static (float Amount, float Speed)[] RadianceCellAbsorptionStats = ItemFactory.CreateCustomSet<(float, float)>((0, 0));
 
         public static int[] ProjectorLensID = ItemFactory.CreateIntSet();
         public static int[] ProjectorLensDust = ItemFactory.CreateIntSet();
@@ -16,9 +19,6 @@ namespace Radiance.Core
         public static Action<ProjectorTileEntity>[] ProjectorLensPreOrderedUpdateFunction = ItemFactory.CreateCustomSet<Action<ProjectorTileEntity>>(null);
         public static Action<ProjectorTileEntity>[] ProjectorLensOrderedUpdateFunction = ItemFactory.CreateCustomSet<Action<ProjectorTileEntity>>(null);
 
-        public static (float Amount, float Speed)[] RadianceCellAbsorptionStats = ItemFactory.CreateCustomSet<(float, float)>((0, 0));
-
-        public static float[] SetPedestalStability = ItemFactory.CreateFloatSet(0);
 
         public static bool[] DrawWindSwayTiles = TilesFactory.CreateBoolSet(false);
